Fixed type

+++ Engine/Core/Utility.cpp
@@ -10,7 +10,7 @@ namespace Engine
         const float diff = max - min;
         value = diff == 0.0f ? 0.5f : (value - min) / diff;
         const float divider = 0.25f;
-        const float number  = floorf(value / divider);
+        const uint8 number  = (uint8)floorf(value / divider);
         const float pos     = (value - number * divider) / divider;
 
         glm::vec3 retVec{0.0f};
